made MainLayout.js
