+++ front-end/src/App.js
@@ -1,34 +1,11 @@
 import './App.css';
-import SideBar from "./components/SideBar";
-import CameraScreen from "./components/CameraScreen";
+import MainLayout from "./components/MainScreen/MainLayout";
 
 function App() {
-  const flexContainer = {
-    display: "flex",
-    backgroundColor: "blue"
-  }
 
-  const sideScreen = {
-    background: "grey",
-    width: "20vw",
-    height: "100vh"
-
-
-  }
-  const cameraScreen = {
-
-    background: "green",
-    width: "80vw"
-
-  }
   return (
-    <div style={flexContainer}>
-      <div style={sideScreen}>
-        <SideBar/>
-      </div>
-      <div style={cameraScreen}>
-        <CameraScreen />
-      </div>
+    <div>
+      <MainLayout />
     </div>
   );
 }
