+++ front-end/src/components/MainScreen/CameraScreen.js
